% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_functions.R
\name{create_bar_chart_levels_ts}
\alias{create_bar_chart_levels_ts}
\title{Generate bar charts for all categorical variables}
\usage{
create_bar_chart_levels_ts(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
trelliscope interactive visualization
}
\description{
Generate bar charts for all categorical variables
}
