% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_functions.R
\name{create_columns_summary_table}
\alias{create_columns_summary_table}
\title{Generate summary table for each column in tibble}
\usage{
create_columns_summary_table(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
datatable
}
\description{
Generate summary table for each column in tibble
}
