% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_functions.R
\name{create_hist_from_numeric}
\alias{create_hist_from_numeric}
\title{Generate histogram from all numeric columns of data}
\usage{
create_hist_from_numeric(data, bin_width = 0.25, num_bins = 30)
}
\arguments{
\item{data}{A tibble}
}
\value{
ggplot2 histogram
}
\description{
Generate histogram from all numeric columns of data
}
