% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_functions.R
\name{create_sampled_datatable}
\alias{create_sampled_datatable}
\title{Create table out of 1000 samples of data from tibble}
\usage{
create_sampled_datatable(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
datatable
}
\description{
Create table out of 1000 samples of data from tibble
}
