% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_functions.R
\name{create_time_series}
\alias{create_time_series}
\title{Generate histogram from all numeric columns of data}
\usage{
create_time_series(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
ggplot2 histogram
}
\description{
Generate histogram from all numeric columns of data
}
