% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_functions.R
\name{create_wordcloud}
\alias{create_wordcloud}
\title{Generate wordcloud from all character data type columns}
\usage{
create_wordcloud(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
wordcloud
}
\description{
Generate wordcloud from all character data type columns
}
