% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_private.R
\name{.is_categorical_col}
\alias{.is_categorical_col}
\title{Checks column data type for either character or factor}
\usage{
.is_categorical_col(col)
}
\arguments{
\item{col}{a column of data (vector)}
}
\value{
Boolean
}
\description{
Checks column data type for either character or factor
}
