% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_private.R
\name{.is_date_time_col}
\alias{.is_date_time_col}
\title{Determines if a column is a dates represented as characters or merely characters}
\usage{
.is_date_time_col(col)
}
\arguments{
\item{col}{a column of data (vector)}
}
\value{
Boolean
}
\description{
Determines if a column is a dates represented as characters or merely characters
}
