% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{.read_data}
\alias{.read_data}
\title{Read in any data source and ensure it is a tibble
PRIVATE FUNCTION}
\usage{
.read_data(filename)
}
\arguments{
\item{filename}{the name and path to the file being ingested into Visualizr}
}
\value{
A tibble of the data
}
\description{
Read in any data source and ensure it is a tibble
PRIVATE FUNCTION
}
\examples{
visualize <- function(filename = "my_file.txt") {
  data <- .read_data(filename)
}
}
\seealso{
visualize
}
