% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_private.R
\name{.select_non_id_columns}
\alias{.select_non_id_columns}
\title{In some cases, it is important to exclude ID variables from datasets. .select_non_id_columns()
takes a tibble and selects only column names which do not mention as being IDs.}
\usage{
.select_non_id_columns(data)
}
\arguments{
\item{data}{a tibble}
}
\value{
a tibble
}
\description{
In some cases, it is important to exclude ID variables from datasets. .select_non_id_columns()
takes a tibble and selects only column names which do not mention as being IDs.
}
