% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_private.R
\name{.split_categorical_cols}
\alias{.split_categorical_cols}
\title{Splits categorical columns into groups of 3 (by default)}
\usage{
.split_categorical_cols(data, n = 3)
}
\arguments{
\item{data}{a tibble}
}
\value{
a vector of grouped categorical column names
}
\description{
Splits categorical columns into groups of 3 (by default)
}
