% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_public.R
\name{gather_group_by_count}
\alias{gather_group_by_count}
\title{Gather, group_by the generated key and value, count, then slice
to create a frequency table of the tibble. Functions with multiple
variables.}
\usage{
gather_group_by_count(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble
}
\description{
Gather, group_by the generated key and value, count, then slice
to create a frequency table of the tibble. Functions with multiple
variables.
}
