% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Generate Rmarkdown report and Rmarkdown standalone file from a single command}
\usage{
visualize(rscript_path = "R/visualize_functions.R",
  new_filename = "test.Rmd")
}
\arguments{
\item{rscript_path}{path to the rscript for Rmd conversion and redending}

\item{new_filename}{path to new Rmd file being created}
}
\description{
Generate Rmarkdown report and Rmarkdown standalone file from a single command
}
