% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Generate Rmarkdown report and Rmarkdown standalone file from a single command.
This function is a wrapper around \code{RParser}, a Python library developed
for parsing R scripts into Rmarkdown.}
\usage{
visualize(new_filename = "./test.Rmd")
}
\arguments{
\item{new_filename}{path to new Rmd file being created}
}
\description{
Generate Rmarkdown report and Rmarkdown standalone file from a single command.
This function is a wrapper around \code{RParser}, a Python library developed
for parsing R scripts into Rmarkdown.
}
