% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Read in a tibble and generate summary statistics and visualizations}
\usage{
visualize(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
RMarkdown file with finished visualizations and summary statistics
}
\description{
Read in a tibble and generate summary statistics and visualizations
}
\examples{
visualize(tbl)
}
